Разделение тестов ProjectEuler на два файла
возможно, слишком большой файл вызывал отказ компилятора
