+++ tests/ProjectEuler/projectEuler.cpp
@@ -20,13 +20,11 @@ SAGA -- это свободной программное обеспечение:
 #include <catch/catch.hpp>
 
 // Используемые возможности
-#include <saga/algorithm.hpp>
 #include <saga/cursor/enumerate.hpp>
 #include <saga/cursor/filter.hpp>
 #include <saga/cursor/iota.hpp>
 #include <saga/cursor/set_union.hpp>
 #include <saga/cursor/stride.hpp>
-#include <saga/cursor/subrange.hpp>
 #include <saga/cursor/take_while.hpp>
 #include <saga/cursor/transform.hpp>
 #include <saga/pipes/filter.hpp>
@@ -1396,95 +1394,6 @@ TEST_CASE("PE 017")
     CHECK(::projectEuler_017(1000) == 21124);
 }
 
-// PE 018: Путь наибольшей суммы (часть I)
-namespace
-{
-    static std::string pe018_sample_data{
-        "3\n"
-        "7 4\n"
-        "2 4 6\n"
-        "8 5 9 3"};
-
-    static std::string pe018_data{
-        "75\n"
-        "95 64\n"
-        "17 47 82\n"
-        "18 35 87 10\n"
-        "20 04 82 47 65\n"
-        "19 01 23 75 03 34\n"
-        "88 02 77 73 07 63 67\n"
-        "99 65 04 28 06 16 70 92\n"
-        "41 41 26 56 83 40 80 70 33\n"
-        "41 48 72 33 47 32 37 16 94 29\n"
-        "53 71 44 65 25 43 91 52 97 51 14\n"
-        "70 11 33 28 77 73 17 78 39 68 17 57\n"
-        "91 71 52 38 17 14 91 43 58 50 27 29 48\n"
-        "63 66 04 68 89 53 67 30 73 16 69 87 40 31\n"
-        "04 62 98 27 23 09 70 98 73 93 38 53 60 04 23"};
-
-    template <class Vector>
-    void projectEuler_018_update(Vector & result, Vector & row)
-    {
-        if(row.size() > 1)
-        {
-            row.front() += result.front();
-            row.back() += result.back();
-
-            for(auto index : saga::cursor::indices(1u, result.size()))
-            {
-                row[index] += std::max(result[index-1], result[index]);
-            }
-        }
-
-        result.swap(row);
-    }
-
-    template <class IntType>
-    IntType projectEuler_018(std::istream & src_in)
-    {
-        std::vector<IntType> result;
-
-        std::string line;
-        std::vector<IntType> row;
-
-        for(; src_in;)
-        {
-            std::getline(src_in, line);
-
-            if(!src_in)
-            {
-                break;
-            }
-
-            std::istringstream line_in(line);
-
-            row.clear();
-            saga::copy(saga::make_istream_cursor<IntType>(line_in), saga::back_inserter(row));
-
-            ::projectEuler_018_update(result, row);
-        }
-
-        assert(!result.empty());
-
-        return saga::max_element(saga::cursor::all(result)).front();
-    }
-
-
-    template <class IntType>
-    IntType projectEuler_018_string(std::string const & text)
-    {
-        std::istringstream istr(text);
-
-        return ::projectEuler_018<IntType>(istr);
-    }
-}
-
-TEST_CASE("PE 018")
-{
-    REQUIRE(projectEuler_018_string<int>(pe018_sample_data) == 23);
-    REQUIRE(projectEuler_018_string<int>(pe018_data) == 1074);
-}
-
 // PE 019 Подсчёт числа воскресений
 namespace
 {
@@ -2008,14 +1917,4 @@ TEST_CASE("PE 030")
 
     REQUIRE(::PE_030(4) == 19316);
     REQUIRE(::PE_030(5) == 443839);
-}
-
-// PE 067: Путь наибольшей суммы (часть II)
-TEST_CASE("PE 067")
-{
-    std::ifstream file("ProjectEuler/p067_triangle.txt");
-
-    REQUIRE(!!file);
-
-    REQUIRE(::projectEuler_018<int>(file) == 7273);
 }
